% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cumulative_exceedance_graphs_method.R
\name{cumulative_exceedance_graphs}
\alias{cumulative_exceedance_graphs}
\title{plot cumulative and exceedance graphs}
\value{
cumulative graph or exceedance graph. the default is cumulative graph
}
\description{
\code{cumulative_exceedance_graphs}
Enables plotting cumulative and exceedance graphs given a climate object
}
\author{
Fanuel Omondi Otieno 2015 (AMI)
}

