% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/histogram_method.R
\name{histogram}
\alias{histogram}
\title{Get histogram from a dataframe}
\value{
It returns a histogram plot.
}
\description{
Allows plotting the data in a histogram plot.
}
\author{
Rafael Posada 2015 (SASSCAL/DWD)
}

