% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeseries_method.R
\name{timeseries}
\alias{timeseries}
\title{Get timeseries from the Climsoft data}
\value{
It returns a timeseries plot.
}
\description{
Allows plotting the data in a timeseries plot.
}
\author{
Rafael Posada 2015 (SASSCAL/DWD)
}

