% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics_frontend_new.R, R/typeWR_frontend.R
\name{typeWR}
\alias{typeWR}
\title{Selection of Windrose type}
\usage{
typeWR()

typeWR()
}
\description{
This is needed for the windrose plots, to select which kind
 of plot is desired to be displayed.

This is needed for the windrose plots, to select which kind
 of plot is desired to be displayed.
}
\author{
Jens Riede, 2015 (SASSCAL/DWD)

Jens Riede, 2015 (SASSCAL/DWD)
}

