% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/windrose_method.R
\name{windrose}
\alias{windrose}
\title{Windrose plots with the openair package}
\arguments{
\item{data_list}{list.}

\item{WR.type}{character. Type of windrose to plot. It can be either
"single", "weekday", "season", "year". Default value: "single"}
}
\value{
Windrose plot
}
\description{
\code{get.windrose}
Creates windrose given a climate object
}
\examples{
climObj <- climate(data_tables=list(data),date_format="\%Y-\%m-\%d")
# where "data" is a data.frame containing the desired data to be plotted.
climObj$get.windrose()
}
\author{
Jens Riede 2015 (SASSCAL/DWD)
}

