% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics_frontend.R
\name{graphics_frontend}
\alias{graphics_frontend}
\title{graphics Frontend}
\usage{
graphics_frontend(climObj)
}
\description{
Allows the interaction with R using "Tcl/Tk" language (see
package \code{tcltk})
}
\author{
Rafael Posada, 2015 (SASSCAL/DWD)
}

