% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics_frontend.R
\name{typeWR}
\alias{typeWR}
\title{Selection of Windrose type}
\usage{
typeWR()
}
\description{
This is needed for the windrose plots, to select which kind
 of plot is desired to be displayed.
}
\author{
Jens Riede, 2015 (SASSCAL/DWD)
}

